feat:添加接口登录校验

+++ src/controllers/todo.controller.ts
@@ -4,7 +4,7 @@ import { StatusCodes } from "http-status-codes";
 import type { ApiResponse } from "@/types/ApiResponse";
 
 /// Local imports
-import { TodoServices } from "../services/todo.services";
+import { TodoServices } from "../services/todo.service";
 import { TodosRoutes } from "@/enums/router.enum";
 import { Variables } from "@/middlewares/todo.di";
 import { log } from "console";
@@ -15,14 +15,23 @@ export class TodoController {
   async createTodo(c: Context<{ Variables: Variables }, TodosRoutes.CREATE>) {
     /// Grab fields
     const { description, isCompleted } = await c.req.json();
+    const userId = c.get('userId');
+    
     log(
       "Creating todo with description:",
       description,
       "and isCompleted:",
       isCompleted,
+      "for user:",
+      userId
     );
+    
     /// Save into database
-    const { id } = await this.todoService.create({ description, isCompleted });
+    const { id } = await this.todoService.create({ 
+      description, 
+      is_completed:isCompleted,
+      user_id:userId 
+    });
 
     /// Response
     return c.json<ApiResponse<{ id: number }>>(
@@ -49,6 +58,19 @@ export class TodoController {
         404,
       );
     }
+    
+    // 验证todo是否属于当前用户
+    const userId = c.get('userId');
+    if (todo.user_id !== userId) {
+      return c.json<ApiResponse>(
+        {
+          success: false,
+          error: "Unauthorized - This todo does not belong to you",
+        },
+        403,
+      );
+    }
+    
     /// Response
     return c.json<ApiResponse<typeof todo>>({
       success: true,
@@ -57,8 +79,12 @@ export class TodoController {
   }
 
   async getTodos(c: Context<{ Variables: Variables }, TodosRoutes.GET_ALL>) {
-    /// Get all todos
-    const todos = await this.todoService.getTodos();
+    log('Getting todos for user:', c);
+    log('Getting todos for user:', c.get('userId'));
+    /// 获取当前用户的所有todos
+    const userId = c.get('userId');
+    const todos = await this.todoService.getTodosByUserId(userId);
+    
     /// Response
     return c.json<ApiResponse<typeof todos>>({
       success: true,
@@ -80,6 +106,19 @@ export class TodoController {
         404,
       );
     }
+    
+    // 验证todo是否属于当前用户
+    const userId = c.get('userId');
+    if (todo.user_id !== userId) {
+      return c.json<ApiResponse>(
+        {
+          success: false,
+          error: "Unauthorized - This todo does not belong to you",
+        },
+        403,
+      );
+    }
+    
     /// Delete todo by id
     await this.todoService.deleteTodoById(Number(id));
     /// Response
